requirements.txt
pyserial>=3.5
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.0.0
